The rc.local file starts the gui automatically.
This is accomplished by running the following commands immediately after booting up:
 cd ComOrgProject/
python gui.py

This file needs to be placed in the /etc/ directory of the pi. The change can also be made directly to your own pi's rc.local file, with the following commands:

sudo nano /etc/rc.local

Then add the commands you want to the file before the "exit 0" line of the code.